requirements.txt
python-fasthtml
mkdocs
mkdocs-material
pymdown-extensions
